server/requirements.txt
Flask
Flask-RESTful
python-dotenv
flask-cors